24-flask-feedback/requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
WTForms==2.3.3
